+++ src/hooks/useChat.tsx
@@ -1,0 +1,8 @@
+import { useContext } from 'react';
+import { ChatContext } from '@/type/ChatType';
+
+export const useChat = () => {
+  const context = useContext(ChatContext);
+  if (!context) throw new Error('ChatProvider 사용 필요');
+  return context;
+};
